+++ .github/workflows/build.yaml
@@ -28,5 +28,6 @@ jobs:
         run: |
           pipenv install pytest
           pipenv run python -m pytest tests/
-
+        env:
+          PYTHONPATH: ${{ github.workspace }}
