+++ fortune_teller/get_fortune.py
